change to int

+++ models/staging/tpch/stg_tpch_customers.sql
@@ -13,7 +13,7 @@ cleanup as (
         c_address as address, 
         c_nationkey as nation_key,
         c_phone as phone_number,
-        c_acctbal as account_balance,
+        c_acctbal::int as account_balance,
         c_mktsegment as market_segment,
         c_comment as comment
